+++ forgerock-auth/src/main/java/org/forgerock/android/auth/Node.java
@@ -11,6 +11,7 @@ import android.content.Context;
 
 import androidx.annotation.VisibleForTesting;
 
+import org.forgerock.android.auth.callback.Callback;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
