+++ forgerock-auth/src/main/java/org/forgerock/android/auth/callback/DerivableCallback.java
@@ -9,8 +9,6 @@ package org.forgerock.android.auth.callback;
 
 import androidx.annotation.Nullable;
 
-import org.forgerock.android.auth.Callback;
-
 /**
  * Interface for {@link Callback} that can be derived.
  */
