+++ ping-protect/src/main/java/org/forgerock/android/auth/callback/PingOneProtectEvaluationCallback.kt
@@ -8,9 +8,13 @@ package org.forgerock.android.auth.callback
 
 import android.content.Context
 import androidx.annotation.Keep
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import org.forgerock.android.auth.FRListener
+import org.forgerock.android.auth.Listener
 import org.forgerock.android.auth.Logger
 import org.forgerock.android.auth.PingOneProtect
-import org.forgerock.android.auth.RootAbstractCallback
 import org.json.JSONObject
 
 private val TAG = PingOneProtectEvaluationCallback::class.java.simpleName
@@ -18,7 +22,7 @@ private val TAG = PingOneProtectEvaluationCallback::class.java.simpleName
 /**
  * Callback to collect the device binding information
  */
-open class PingOneProtectEvaluationCallback : RootAbstractCallback {
+open class PingOneProtectEvaluationCallback : AbstractCallback {
 
     @Keep
     constructor(jsonObject: JSONObject, index: Int) : super(jsonObject, index)
@@ -59,23 +63,24 @@ open class PingOneProtectEvaluationCallback : RootAbstractCallback {
         super.setValue(value, 1)
     }
 
-//    /**
-//     * Request for Integrity Token from Google SDK
-//     *
-//     * @param context  The Application Context
-//     * @param listener The Listener to listen for the result
-//     */
-//    open fun getSignals(context: Context,
-//                        listener: FRListener<Void>) {
-//        val scope = CoroutineScope(Dispatchers.Default)
-//        scope.launch {
-//            try {
-//                Listener.onSuccess(listener, null)
-//            } catch (e: Exception) {
-//                Listener.onException(listener, e)
-//            }
-//        }
-//    }
+    /**
+     * Request for Integrity Token from Google SDK
+     *
+     * @param context  The Application Context
+     * @param listener The Listener to listen for the result
+     */
+    open fun getSignals(context: Context,
+                        listener: FRListener<Void>
+    ) {
+        val scope = CoroutineScope(Dispatchers.Default)
+        scope.launch {
+            try {
+                Listener.onSuccess(listener, null)
+            } catch (e: Exception) {
+                Listener.onException(listener, e)
+            }
+        }
+    }
 
     open suspend fun getSignals(context: Context) {
         try {
